// src/client/blog/api.ts
export interface CategoryPublic {
  /**
   *
   * @type {string}
   * @memberof CategoryPublic
   */
  _id: string
  /**
   *
   * @type {string}
   * @memberof CategoryPublic
   */
  name: string
  /**
   *
   * @type {Array<SubcategoryPublic>}
   * @memberof CategoryPublic
   */
  subcategories?: Array<SubcategoryPublic>
}
// ...
export interface PostPublic {
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  _id: string
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  title: string
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  content: string
  /**
   *
   * @type {boolean}
   * @memberof PostPublic
   */
  is_public: boolean
  /**
   *
   * @type {CategoryPublic}
   * @memberof PostPublic
   */
  category?: CategoryPublic | null
  /**
   *
   * @type {SubcategoryPublic}
   * @memberof PostPublic
   */
  subcategory?: SubcategoryPublic | null
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  author: string
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  created_at?: string | null
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  updated_at?: string | null
}
// ...
export interface SubcategoryPublic {
  /**
   *
   * @type {string}
   * @memberof SubcategoryPublic
   */
  _id: string
  /**
   *
   * @type {string}
   * @memberof SubcategoryPublic
   */
  name: string
}